+++ packages/glsp-client/src/browser/crossmodel-diagram-module.ts
@@ -2,7 +2,7 @@
  * Copyright (c) 2024 CrossBreeze.
  ********************************************************************************/
 
-import { GRID, ShowPropertiesAction } from '@crossmodel/protocol';
+import { ExpandNavigatorForNewFileAction, GRID, ShowPropertiesAction } from '@crossmodel/protocol';
 import {
    ConsoleLogger,
    GLSPHiddenBoundsUpdater,
@@ -36,6 +36,7 @@ import { CrossModelDiagramStartup } from './cross-model-diagram-startup';
 import { CrossModelErrorExtension } from './cross-model-error-extension';
 import { CrossModelToolPalette } from './cross-model-tool-palette';
 import { CrossModelGLSPSelectionDataService } from './crossmodel-selection-data-service';
+import { ExpandNavigatorActionHandler } from './expand-navigator-action-handler';
 import { ShowPropertiesActionHandler } from './show-properties-action-handler';
 
 export function createCrossModelDiagramModule(registry: interfaces.ContainerModuleCallBack): ContainerModule {
@@ -69,6 +70,7 @@ export function createCrossModelDiagramModule(registry: interfaces.ContainerModu
       rebind(GLSPHiddenBoundsUpdater).to(CrossModelHiddenBoundsUpdater).inSingletonScope();
 
       configureActionHandler(context, ShowPropertiesAction.KIND, ShowPropertiesActionHandler);
+      configureActionHandler(context, ExpandNavigatorForNewFileAction.KIND, ExpandNavigatorActionHandler);
    });
 }